Add rotating cube gif to README and implement game tick interval
A gif image showing a rotating cube has been added to the README file for better demonstration of the project. In the main.kt changes have been implemented to manage the game loop by introducing an UPDATE_INTERVAL check which ensures the 'render' and 'present' methods are called at a set pace rather than as fast as possible.