Update README and modify function declarations
The README file was updated with instructions on how to run the project and a new screenshot section. In the Kotlin code, changes were made to the 'initialize' function in 'RotatingCubeScene.kt' and 'createScene' function in 'render.kt', removing the 'suspend' keyword from their declaration. This may impact how these functions are called or how concurrency is managed.