Add support for WebAssembly (WASM) in the project
This commit introduces changes to the build configuration and README, thus adding support for a new WASM target. Alongside the change, some array declaration got converted to list in the code base, and a new HTML file for the WASM execution environment has been added. The wgpu4k version in gradle has also been updated to support the changes.